+++ bootstrap.js
@@ -21,7 +21,10 @@ var options = {
 	scratch: path.resolve(PATH_TEMP, 'scratch'),
 
 	// The cache map for resized images & thumbnails
-	cache: path.resolve(PATH_TEMP, 'imagecache')
+	cache: path.resolve(PATH_TEMP, 'imagecache'),
+
+	// Path to fontawesome pro
+	fontawesome_pro: null,
 };
 
 // Inject the user-overridden options
@@ -65,6 +68,10 @@ options.veronica = new Veronica({
 });
 
 alchemy.hawkejs.addRawHtml(`<script>document.cookie = 'mediaResolution=' + encodeURIComponent(JSON.stringify({width:Math.max(screen.availWidth||0, window.outerWidth||0) || 1024,height:Math.max(screen.availHeight||0, window.outerHeight||0) || 768,dpr:window.devicePixelRatio}))</script>`);
+
+if (options.fontawesome_pro) {
+	alchemy.exposeStatic('fontawesome_pro', options.fontawesome_pro);
+}
 
 /**
  * Register a profile under the given name